ClassFileWriter no longer performs verification
Previously, the ClassFileWriter performed verification of the class file
it was to write.  This worked well, but it was also rather inflexible
from a debugging perspective.  Specifically, because the verifier does
not report particularly helpful error messages.  Therefore, I've stopped
the ClassFileWriter from doing this and becomes up to clients to take
care of this (if they wish).  This makes it easier to debug as they can
enable / disable it as they see fit.